(PC-26940)[API] chore: remove one time used mixin
